requirements.txt
Django==3.2.15
adal==1.2.7
aiohttp==3.7.4
aioredis==1.3.1
amqp==5.0.6
APScheduler==3.8.0
argcomplete==1.10.0
asgiref==3.4.1
asn1crypto==1.4.0
async-timeout==3.0.1
attrs==21.2.0
autobahn==20.12.3
Automat==20.2.0
beautifulsoup4==4.8.0
billiard==3.6.4.0
blis==0.7.4
botbuilder-core==4.14.0
botbuilder-schema==4.14.0
botframework-connector==4.14.0
cachetools==4.2.2
catalogue==2.0.6
certifi==2022.12.7
cffi==1.14.6
channels===3.0.4
channels-redis==3.3.0
chardet
click==7.1.2
constantly==15.1.0
cryptography==3.4.8
cymem==2.0.5
daphne==3.0.2
decorator==4.4.2
defusedxml==0.7.1
django-ace==1.0.11
django-cors-headers==3.9.0
django-crontab==0.7.1
django-oauth-toolkit==1.5.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-session-timeout==0.1.0
django-suit-redactor==0.0.4
django-user-agents==0.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-xml==2.0.0
docopt==0.6.2
docx2txt==0.8
EbookLib==0.17.1
editdistance==0.5.3
elasticsearch==7.15.0
emoji==2.2.0
entrypoints==0.3
extract-msg==0.28.0
fasteners==0.16.3
ffmpeg-python==0.2.0
flake8==3.9.2
flake8-polyfill==1.0.2
future==0.18.2
fuzzywuzzy==0.18.0
geocoder==1.38.1
gevent==21.8.0
google-api-python-client==2.23.0
google-apitools==0.5.32
google-auth==2.2.1
google-auth-httplib2==0.1.0
google-businesscommunications==1.0.4
google-businessmessages==1.0.4
googletrans==3.0.0
greenlet==1.1.1
gunicorn==20.1.0
hiredis==2.0.0
html5lib==1.1
httplib2==0.19.1
hyperlink==21.0.0
idna==2.10
idna-ssl==1.1.0
IMAPClient==2.1.0
imgkit==1.2.2
importlib-metadata==4.8.1
incremental==21.3.0
isodate==0.6.0
jsonpickle==1.4.2
kombu==5.1.0
lxml==4.9.1
mccabe==0.6.1
msal==1.6.0
msgpack==1.0.2
msrest==0.6.10
multidict==5.1.0
murmurhash==1.0.5
nltk==3.6.6
num2words==0.5.10
numpy==1.22.0
oauth2client==4.1.3
oauthlib==3.2.1
olefile==0.46
orderedset==2.0.3
pandas==1.3.3
pdfminer.six==20200517
pep8-naming==0.12.1
Pillow==9.3.0
pkcs7==0.1.2
plac==1.3.3
preshed==3.0.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
PyExecJS==1.5.1
pyflakes==2.3.1
PyHamcrest==2.0.2
pyinflect==0.5.1
pyOpenSSL==21.0.0
pysaml2==7.0.1
pyspellchecker==0.6.2
python-dateutil==2.8.2
python-docx==0.8.11
python-pptx==0.6.21
python-slugify==5.0.2
pytz==2021.1
ratelim==0.1.6
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
rsa==4.7.2
huggingface-hub==0.10.1
sentence-transformers==2.2.0
service-identity==21.1.0
singledispatch==3.7.0
six==1.16.0
sortedcontainers==2.4.0 
soupsieve==2.2.1
SpeechRecognition==3.8.1
sqlparse==0.4.2
srsly==2.4.1
text-unidecode==1.3
textblob==0.15.3
thinc==8.0.10
tldextract==3.1.2
tqdm==4.62.3
Twisted==22.4.0
txaio==21.2.1
typing-extensions==3.10.0.2
tzlocal==2.1
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.26.7
user-agents==2.2.0
vine==5.0.0
virtualenv==20.8.1
wasabi==0.8.2
webencodings==0.5.1
websocket-client==1.2.1
whitenoise==5.3.0
XlsxWriter==3.0.1
xlwt==1.3.0
xmltodict==0.12.0
yarl==1.6.3
zipp==3.5.1
zope.interface==5.4.0
moviepy==1.0.3
dict2xml==1.7.0
google-cloud==0.34.0
google-cloud-core==2.0.0
google-cloud-translate==3.4.0
python-magic==0.4.24
openpyxl==3.0.9
wordninja==2.0.0
croniter==1.0.15
plotly==5.3.1
kaleido==0.2.1
xlrd==1.2.0
spacy==3.1.3
pyenchant==3.2.1
scikit-learn==1.0
sklearn==0.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
textract
pdfplumber==0.5.28
func-timeout
Flask==1.0.0
flask_api==1.0
google-cloud-pubsub
boto3==1.20.8
demoji==1.1.0
selenium==4.1.0
webdriver-manager==3.5.3
authlib
email-reply-parser==0.5.12
phonenumbers==8.12.46
requests-html==0.10.0
viberbot
pdfkit==1.0.0  
forex-python
kafka-python
